Moved survivor data and logic to survivor component from parent
